feat: Add GET_ALAMAT and ADD_ALAMAT routes for customer addresses
This commit adds the GET_ALAMAT and ADD_ALAMAT routes to the CUSTOMER_ROUTES object in the Routes.js file. The GET_ALAMAT route is used to retrieve customer addresses, while the ADD_ALAMAT route is used to add a new address. These routes are necessary for managing customer addresses and improving the user experience.

Refactor the Checkout.jsx component to include functionality for selecting a delivery option and choosing an address. The component now includes a radio button for selecting between "Ambil di Toko" (Pick up at the store) and "Delivery" options. When the "Delivery" option is selected, a dropdown menu is displayed for selecting an address from the available options. Additionally, a "Tambah Alamat?" (Add Address?) link is provided to open a modal for adding a new address.

These changes enhance the functionality of the application and improve the user experience for customers during the checkout process.